+++ conf/main/grakn.properties
@@ -58,13 +58,6 @@ server.port=4567
 # Set the external folder serving static files to the Grakn dashboard.
 server.static-file-dir=./services/assets/
 
-# Number of times to retry graph mutations if a temporary backend exception occurs
-# during the mutation. If this number is set to 0, the server will not retry.
-loader.repeat-commits=5
-
-# The delay in seconds between receiving a task and executing it.
-# This is to give time for any indices relating to the task to be finalised
-tasks.delay=1800
 
 # This defines the number of threads to be used when executing tasks. Consider
 # increasing this number to increase engine task throughput and CPU usage.
@@ -118,5 +111,5 @@ cache.db-cache-size=0.25
 ############################# Redis Configuration #############################
 # IP or hostname on which Redis is listening for connections.
 queue.host=localhost:6379
-queue.consumers=32
+post-processor.pool-size=32
 redis.pool-size=32
