+++ conf/test/tinker/grakn.properties
@@ -30,12 +30,10 @@ server.host=0.0.0.0
 server.static-file-dir=./services/assets/
 webserver.threads=64
 
-#RESTLoader Config
-loader.repeat-commits=5
 
 #Redis Config
 queue.host=localhost:6379
-queue.consumers=32
+post-processor.pool-size=32
 redis.pool-size=32
 
 ####################################
@@ -57,5 +55,3 @@ loader.threads=1
 # whether to start an embedded engine, cassandra and redis for tests
 test.start.embedded.components=true
 
-# Task execution delay in seconds
-tasks.delay=1
