+++ grakn-engine/src/main/java/ai/grakn/engine/GraknCreator.java
@@ -22,11 +22,11 @@ import ai.grakn.engine.data.RedisWrapper;
 import ai.grakn.engine.factory.EngineGraknTxFactory;
 import ai.grakn.engine.lock.JedisLockProvider;
 import ai.grakn.engine.lock.LockProvider;
-import ai.grakn.engine.postprocessing.IndexPostProcessor;
-import ai.grakn.engine.postprocessing.InstanceCountPostProcessor;
-import ai.grakn.engine.postprocessing.PostProcessor;
-import ai.grakn.engine.postprocessing.RedisCountStorage;
-import ai.grakn.engine.postprocessing.RedisIndexStorage;
+import ai.grakn.engine.task.postprocessing.IndexPostProcessor;
+import ai.grakn.engine.task.postprocessing.InstanceCountPostProcessor;
+import ai.grakn.engine.task.postprocessing.PostProcessor;
+import ai.grakn.engine.task.postprocessing.RedisCountStorage;
+import ai.grakn.engine.task.postprocessing.RedisIndexStorage;
 import ai.grakn.engine.util.EngineID;
 import com.codahale.metrics.MetricRegistry;
 import redis.clients.jedis.Jedis;
