+++ grakn-test/test-integration/src/test/java/ai/grakn/test/engine/postprocessing/PostProcessorTest.java
@@ -25,8 +25,8 @@ import ai.grakn.Keyspace;
 import ai.grakn.concept.Concept;
 import ai.grakn.concept.ConceptId;
 import ai.grakn.concept.EntityType;
-import ai.grakn.engine.postprocessing.PostProcessor;
-import ai.grakn.engine.postprocessing.RedisCountStorage;
+import ai.grakn.engine.task.postprocessing.PostProcessor;
+import ai.grakn.engine.task.postprocessing.RedisCountStorage;
 import ai.grakn.kb.log.CommitLog;
 import ai.grakn.test.rule.EngineContext;
 import ai.grakn.util.SampleKBLoader;
